pdb files are first class citizens on windows
Summary:
On windows debug information is stored in pdb files. Those files are produced if linker is invoked with `/DEBUG` flag - a corresponding pdb file is placed near its "parent" binary file (shared lib - dll - is also a binary file). This is a common practice in msvc world just to use `/DEBUG` and rely on the fact that for a `lib.dll` the pdb file will be `lib.pdb` and for an `app.exe` it will be `app.pdb`. Pdb files should be cached.

In order to handle this, the concept of derived extra outputs is introduced. Basically, if `WindowsLinker` detects the `/DEBUG` flag then the corresponding pdb file is handled as an extra output, so it will be cached and also can be referenced via `$(location :target[pdb])`.

Reviewed By: cjhopman

fbshipit-source-id: 4ddbe83