Add hashCode hashing and improved equals on ImmutableMapWithNullValues
Summary:
In Guava 25, ImmutableSet building has changed such that we compute
the hashcode and equals during add, instead of build. This greatly improves
the perf of ImmutableSet when hashing and equals are not slow.

However, we are adding ImmutableMapWithNullValues into ImmutableSets many times,
which caused a significant regression due to heavy computation of the hashcode
and equals on the maps.

This caches it.

Reviewed By: styurin

fbshipit-source-id: 2616b9a