Avoid copying resources reachable via exported deps of header-only dependencies.
Summary:
Resources reachable via exported deps of header-only (and other
non-static-linked) dependencies were being copied into the bundle.

I believe this was specified unintentially and is incorrect. The exported-deps
behavior for linking exists due to headers referencing exported symbols from a
dependency, it should not apply to resource copying.

There is a legacy behavior where a library can depend on a framework as an "exported_dep", which will cause that framework's resources to be copied to the bundle containing the library. I don't think this is useful behavior so I removed the test for this.

Reviewed By: ttsugriy

fbshipit-source-id: e911afc