Remove searching for Swift 2.3 toolchain (#1907)
Summary:
Xcode 8.0 - 8.2.1 supported switching swift versions by having an extra `Swift_2_3.toolchain` embedded in Xcode, and overriding the search paths for tools like swiftc if the version was changed.

Since then Xcode/Swift's model for supporting different versions of the swift source by passing in a `-swift-version` flag to the compiler. This was added to buck, so we now pass this flag instead of overriding toolchains. As earlier swift toolchian versions won't understand the flag which is now the buck way of specifying a language, this toolchain search code can now be cleaned up.
Closes https://github.com/facebook/buck/pull/1907

Reviewed By: milend

Pulled By: milend

fbshipit-source-id: 3a81eaf