Allow option for haskell_ghci to generate multiple scripts
Summary:
Take in a list of extra_templates to generate scripts as part
of `haskell_ghci` TARGET. This would allow using the haskell compiler
to do other stuff (like typechecking) in addition to using it as a REPL tool
(ghci).
NOTE : These scripts are currently generated in the top level
`buck-out/path/to/target` directory, irrespective of the relative path
of the templates themselves. Therefore, a template at `xyz-directory/template_file` would not create any `xyz-directory` in the buck-out directory for the generated script. No particular reason for this - seemed simple and sufficient for the use-case.

Reviewed By: cjhopman

fbshipit-source-id: 5695f4f