fix paths for source files being in different locations
Summary:
cgo tool assumes that all the files are in source dir and does some magic with replacing paths with '_'. This happen to be problematic with genrules or even if the source file is in directory like (dir/dir/file.go).

However with -srcdir flag passing the filepaths works much better and fixes the issue of long paths (later on translated to pathnames)

fbshipit-source-id: f2ef786