Add `buck audit mbr isolation <targets>`
Summary:
This generates an "isolation report" which includes a bunch of
information about which nodes in the action graph can be sent to be
built by an isolated service (i.e. remote execution).

It serializes every node in the graph and reports exceptions thrown
(typically due to some referenced object not supporting serialization)
and every absolute path that is seen.

It does the same for computing inputs, and in addition it follows
symlinks/directory contents (just like we do when we actually have to
send inputs off to an isolated service).

It has builtin caching (similar to rulekey computation) so that each
AddsToRuleKey instance is only processed once.

`buck audit mbr isolation buck`: https://gist.github.com/cjhopman/a7b9cd5baa5f612ccad664e80429fff1

Reviewed By: ttsugriy

fbshipit-source-id: 3219479