Use PrefixMapDebugPathSanitizer to sanitize assembled object files produced in Android NDK build
Summary:
Android NDK 16 started to generate debug symbols in object files assembled from asm files when using gcc for x86_64 (when building on OS X).

`PrefixMapDebugPathSanitizer` does the job here, not need to use `PrefixMapDebugPathSanitizer` at all.

The original change that added this logic was based on the assumption that clang on OS X didn't use `-fdebug-prefix-map` with asm files and as a result object files contained the current directory.

It doesn't seem to be the case with Android NDK. We have a test (`com.facebook.buck.android.toolchain.ndk.impl.NdkCxxPlatformIntegrationTest#testWorkingDirectoryAndNdkHeaderPathsAreSanitized`) which is failing with Android NDK 16 because asm files for x86_64 weren't sanitized properly.

Reviewed By: bobyangyf

fbshipit-source-id: c378d17