Update debian build files
Summary: Update the debian dockerfile and build scripts to accept a build version/timestamp. Notably: This replaces the 'tag' variable with a 'version' one that is not 'v' prefixed, and tells the deb to install to /usr/local/bin/buck, rather than /usr/bin/buck. This is not a distro provided application, so /usr/local/bin is more appropriate.

Reviewed By: ttsugriy

fbshipit-source-id: 2115182