Allow augmentation of internal errors in ErrorLogger
Summary:
Currently a lot of buck's error messages are presented as internal
errors (even when they should logically be user errors), but the
ErrorLogger's augmentor is only applied to user errors.

Long term, it's unclear if we will ever get to a state where
users/integrators will have no use for augmenting internal errors, but
right now it'll definitely be useful.

Reviewed By: philipjameson

fbshipit-source-id: 85a6bd0