Use `rm -f` instead of `rm` for deleting files on device
Summary:
When a file is read-only, Android's shell will ask for user confirmation
to delete the file. That's quite inconvenient when buck is invoking it
and currently just hangs. Use `rm -f` instead. This has the added
benefit that, if the file is owned by root but the directory is writable
by the current user, deleting the file will succeed.

Reviewed By: asp2insp

fbshipit-source-id: 7e24c9a