Add support for creating symlinks in privileged and non-privileged mode.
Summary:
Depending on the level of support in the OS (Windows), make sure the symlink creation works in both cases - when non-privileged CreateSymbolicLink is supported and if not require elevated run of Buck.
Fixes a problem with imported API from Windows. The CreateSymbolicLinkW API on Windows returns a BOOLEAN, which is really a byte in Java (8 bit unsigned numeric value).
At the JVM level, boolean, byte, short, char, and int are all treated as ints, and boolean is indeed represented as 0 and 1. The API is successful if the return value is not 0. So, returning 0xD0 should be a success, but the 0 bit is not set and Java will interpret this as boolean false. This is fixed by making the function imported from native as a byte return and compering the return value explicitly to 0.

Reviewed By: philipjameson

fbshipit-source-id: 075f85d