create gopath/goroot symlinktree
Summary:
buck is very flexible on how go packages are structured. In general it's pretty good feature, but...
1. IDE's are confused on where particular files / packages are located, which makes the IDE integration pretty hard (autocomplete etc.)
2. You're unable to build a project with standard go build command (which many times comes handy)

This diff creates the symlinktree to all the source files used during the build. This creates the GOPATH and GOROOT directories which can be used with
standard go build / go test command or IDEs.

The feature is enabled with the option passed via .buckconfig or the cli

Reviewed By: sbalabanov

fbshipit-source-id: da83a8c