Extract JavaAbiInfo from HasJavaAbi
Summary:
Currently dependent rules are holding direct references to other rules
as HasJavaAbi interfaces. We need to break these direct cross-rule
references. This introduces JavaAbiInfo which contains most of the
functions from HasJavaAbi. HasJavaAbi is updated to expose a
JavaAbiInfo. In this way, during AG construction, we can extract the
JavaAbiInfo and in dependent rules hold only a reference to that.

Reviewed By: styurin

fbshipit-source-id: 2486fbc