Add support for plugin-based commands
Summary:
This change adds generic infrastructure for plugin-based commands.

Plugin-based commands allow to load subcommands using plugin framework and offload some or all of the functionality to a subcommand.

The majority of the changes are dealing with argument parsing.

The arguments are handled in the way similar to the current situation (when a user can provide multiple arguments some of which are mutually exclusive or not used together, for example, IntelliJ-specific options when running `buck project` for XCode). The framework loads subcommands from plugins and then parses all of the arguments and then a specific command implementation is free to do whatever is necessary with those arguments.

Another approach that was considered is to make argument parsing stricter by using an option to select one subcommand (for example, `--ide` in case of `buck project`). This adds a lot of complexity in cases when this option should be derived from configuration (it's possible to set `--ide` in `.buckconfig` to avoid specifying it in the command line) since the command line options can contain configuration options. This approach would also require a specific order of arguments in the command line or major changes in the args4j framework.

Reviewed By: ttsugriy

fbshipit-source-id: 29f3d1c0fd