Add in proguard default config for storing dead code list
Summary:
We need to store `usage.txt` for feature appsize analysis.

`usage.txt` Lists the code that was removed from the APK.
> -printusage [filename]
Specifies to list dead code of the input class files. The list is printed to the standard output or to the given file. For example, you can list the unused code of an application. Only applicable when shrinking.

Reviewed By: styurin

fbshipit-source-id: d4a18d2