Introduce RawTargetNode
Summary:
`RawTargetNode` is a node similar to `TargetNode`, but its attributes are kept in a map instead of an immutable object. This will allow us to keep objects of various types for an attribute (which cannot be easily done with an object created using Immutables.) We will use it to keep selectable values.

Another distinction from `TargetNode` is not keeping a reference to `Description` and using `BuildRuleType` which can be used to find `Description`.

Reviewed By: cjhopman

fbshipit-source-id: e955272f39