Add RawTargetNodeFactory
Summary:
`RawTargetNodeFactory` creates instances of `RawTargetNode` using the data obtained from build file parsers, i.e. non-coerced attributes.

`ConstructorArgMarshaller` was extended to support transforming the raw data coming from build file parsers to a form suitable for `RawTargetNode`. When an attribute value contains Bazel's `SelectorList` it's coerced by a coercer that supports this data type, otherwise the value is coerced using its type. The reason why we need that logic is because `SelectorList` is not generic type, but we need to treat it like one.

Reviewed By: bobyangyf

fbshipit-source-id: 2275140706