Add BuildReport event to Buck
Summary:
This BuildReport event will post if a build that includes the tag "--build-report" is run and has a reportable failure. In .buckconfig this event will be enabled to write to Scribe. It will be used to display build reports on the Buck UX.

The BuildReport constructor can now take an eventBus, which gets passed in where it is available (in BuildCommandTest, I assume that it's ok not to post the event to any bus but I could be wrong).

Reviewed By: sbalabanov

fbshipit-source-id: 10b62a0