rust: add #save-analysis flavour
Summary:
`#save-analysis` is an extension of `#check`. It performs the same actions as a `#check` build, but also
emits a `.json` file containing full type/symbol cross-reference information, for consumption by
tools like `rls` (Rust Language Services).

This needed some refactoring of how `#check` was handled. `#save-analysis` is treated as a variant of `#check`,
but `#check` it was not considered to produce any (useful) output for binaries. However, the
save-analysis output is important for binaries, so I just made check's output also "useful". (This
has the side-effect of allowing the result of successful check builds of binaries to be cached.)

Reviewed By: styurin

fbshipit-source-id: 093a1f7c16