Extend PerBuildStateFactory to create a TargetNode pipeline with configurable attributes
Summary: The most important change is in `PerBuildStateFactory` which is extended to use `TargetNode` pipeline based on `RawTargetNode` if it's enabled in `.buckconfig` (`parser.enable_configurable_attributes`). By default it's using the old `TargetNode` pipeline that produces nodes from raw parse data. The rest of this change in mostly plumbing.

Reviewed By: cjhopman

fbshipit-source-id: 1792ec2925