Add gray borders around code-styled text
Summary:
Add a gray border around code-styled text in order to make it easier to visually distinguish code inside running text.

This is a similar style to that used on Slack. GitHub and Bazel do not use this technique. However, they //shade// their inline code dark-gray against a white background whereas we shade lighter against a darker background, which doesn't seem to produce as much contrast.

Reviewed By: styurin

fbshipit-source-id: 458b8696e8