Add output of prebuilt_jar as class root
Summary:
Given the following scenario:

```
genrule(
  name = 'make_jar',
  cmd = '...'
  out = 'my.jar',
)

prebuilt_jar(
  name = 'my.jar',
  binary_jar = ':make_jar',
)
```

the IJ generated project will point to the output of the genrule. But `prebuilt_jar` will copy the input jar when built, thus producing a cacheable output. So if you get a cache hit on the `prebuilt_jar`, then `buck` will never populate the output location of the `genrule` and now the library iml for IJ points to the wrong location. This diff fixes this by adding both the input and the output paths for a `prebuilt_jar` to the IJ project's library iml.

Reviewed By: styurin

fbshipit-source-id: de223440d6