fix: buck project was removing files generated within the same directory as part of a cleanup + add project_path option
Summary:
The "cleanup" loop was removing files that were generated by the same loop (assuming there are more than one file in the directory).
It's fixed within this diff.

To the 'project_path':
The vendor_path is a list of vendor paths so it's not really valid to use it as a destination for generated files.

Reviewed By: styurin

fbshipit-source-id: a5163c5a98