Sort cells to be independent of order in .buckconfig
Summary:
During cache invalidation, `lookup` tries to match a `Path` to a cache in the `StackedFileHashCache`, but it returns the first cache it matches. When you introduce nested cells, this causes problems because the lookup can return the incorrect cache.

Currently, the order the caches appear in the `StackedFileHashCache` is based off of the ordering of the `repositories` in the buckconfig file. `WatchedFileHashCaches` are ordered the same as the buckconfig, and the `DefaultFileHashCaches` are ordered the same as well, except that the cell root is the first `DefaultFileHashCache`.

This diff sorts `getKnownRoots` so that roots with a longer path appear early in the list of caches if they have the same prefix. That way, when `lookup` is called, it will go check the deepest nested cell first before checking other cells. This helps enforce the idea that buck should associate every path with only one cache (the cache belonging to the closest ancestor with a buckconfig).

Reviewed By: cjhopman

fbshipit-source-id: 60078dbd1e