Fix cache invalidation issue when new flavored node without unflavored node added
Summary: There are cases, e.g. when a `cxx_binary` is changed to a `cxx_library`, where we can go from a target graph with an unflavored node to a target graph with a flavored version of that node without the corresponding unflavored node, where we'd fail to invalidate the unflavored node. This change fixes the invalidation logic to account for this case.

Reviewed By: bobyangyf

fbshipit-source-id: 5265ecc6f0