Disable deprecation warning about importing imp in __test_main__.py
Summary:
When using Python 3+, `import imp` emits a deprecation warning that is shown when a test fails. The messages warns about deprecating of `imp` in favor of `importlib`.

`importlib` was added in Python 3+ and since Buck supports both 2+ and 3+ versions of Python we cannot replace `imp` with `importlib` given that `importlib` doesn't provide functionality identical to `imp`.

This change disables the warning for `import imp` only, not affecting other imports.

Reviewed By: philipjameson

fbshipit-source-id: ca50b61365