Remove buildtarget from shellstep and a bunch of subclasses
Summary:
Most of these uses are just to add the buildtarget to an exception
message, but the CachingBuildEngine already handles adding the build
target as context to any failure.

Having access to a BuildTarget allows a step to construct paths on its
own to things (due to BuildTarget.getBasePath()), but that makes it
difficult to restrict what paths can be created.

Reviewed By: styurin

fbshipit-source-id: dcbd0a3f02