add min number of samples before reporting error
Summary: This allows you to set a minimum amount of samples before determining if a connection is unhealthy. This is to avoid the cases when a connection doesn't get a lot of requests and if one is an error then it forces the connection to be marked unhealthy since it goes from 0% to 100% error ratio.

Reviewed By: shivanker

fbshipit-source-id: 760b1b44c8