Keep the OS lineseparator when capturing output.
Summary: The code that I removed was added in 2013 and it erases the Windows line separator, replacing it with the Unix one. We should always use the underlined OS file separator. The code was added to make a py file works and I am not sure this file is still in use. This PR is also to run the tests for different platforms to make sure things don't break iof the lineseparator is kept as it should be - the one for the underlined OS.

Reviewed By: ttsugriy

fbshipit-source-id: 7659b4917d