genrule: allow specifying sources as a map
Summary:
Similar to the `headers` parameter in `cxx_library`, this updates the `srcs`
parameter of `genrule`s to use a `SourceList`, allowing sources to be specified
as a dict/map.  This allows users to easily change the layout of sources in the
symlink source tree the rule sets up to run under.

Reviewed By: cjhopman

fbshipit-source-id: fc56702bbb