genrule: move `cacheable` to base args class
Summary:
This makes the `cacheable` parameter part of the base genrule args class, so
that all genrule-style rules have it available.  Generally, the `cacheable` parameter
should be avoided, as things that make a particular rule uncacheable (e.g.
non-determinism) also break other parts of the build system (e.g. input rule keys),
and this is only provided for hacky workarounds.

Reviewed By: philipjameson

fbshipit-source-id: 3c8020cb3f