Add ability to handle hashes in test parameterization name
Summary:
JUnit parameterized tests can name their parameters, which can contain
"#"'s. These are special markers used by buck for splitting classname and
methodname, which conflicts.

We fix this by splitting only on the first "#", as class name cannot contain
special characters, and so the first "#" is guaranteed to be the class
method divider character

Reviewed By: ttsugriy

fbshipit-source-id: e422951744