Fix action graph invalidation regressions
Summary: Fixes regressions from the previous `IncrementalActionGraphGenerator` change. Adds logic to (1) invalidate all flavors of a node if a new flavor appears, and (2) avoid pulling in rules from cache for nodes that go away. The latter was causing a "sticky tests" issue, as `TestCommand` looks for all action graph nodes of type `TestRule`.

Reviewed By: bobyangyf

fbshipit-source-id: 57ff256b4a