Make shell use stored path to project root
Summary:
Previously, the shell script depends on a relative path to the project root
which breaks when the buck-out is not within the project-root.
When users configured buck-out location to be different, the relative path would
also break if the remote cache is used.

This adds a new symlink from buck-out to the project root so that shell scripts
can consistently refer to the project root regardless of where the buck-out lives

Reviewed By: philipjameson

fbshipit-source-id: 905747c5f6