Remove strange memoization
Summary:
Currently JarBuildStepsFactory is doing an odd map of target to
outputs so that it can cache the output computation for both the jar and
the abis. This was added because it showed a significant perf boost in
action graph creation.

It's odd for the JBSF to be doing this caching and for it to be stored
in a generic map (when we really know it's just for a handful of
specific targets). It also requires it to have access to the
ProjectFilesystem.

Change this to just cache the sourcepaths for getSourcePathToOutput in
the rules themselves.

Tested on AG creation time of a large app and saw no regression.

Reviewed By: styurin

fbshipit-source-id: c314e57191