Introduce CompilerOutputPaths
Summary:
Extracts the logic related to computing and exposing the various output
paths used by the compiler into a single immutable object.

This no longer allows customizing of any of the previously customizable
paths. This is easier to reason about. In a couple cases where the
control flow made it difficult to use the CompilerOutputPaths path in a
case where it was previously being customized, some Preconditions are
added to ensure that they match up as expected.

Reviewed By: styurin

fbshipit-source-id: e768a5e774