Add RecordArtifactVerifier and use it
Summary:
One big difference between legacy BuildRules and ModernBuildRules is
that legacy rules need to manually record all their outputs but MBRs
will do this automatically.

In some cases (like Java) the artifact recording is happening deep in
the step factories/implementations and it can be difficult to extract
that up to a higher level.

To make migration to MBR easier, this introduces RecordArtifactVerifier.
This will record a fixed set of artifacts and then act as a
BuildableContext that verifies that any further recorded artifacts were
in the fixed list (or a child of an item in the list). Then, a rule can
replace the BuildableContext in getBuildSteps() with this and it ensures
that the steps/lower levels are well-behaved.

Updates DefaultJavaLibrary/CalculateSourceAbi to use this.

Reviewed By: jtorkkola

fbshipit-source-id: 8c2a080cc8