Make AndroidPrebuiltAar a little simpler
Summary:
AndroidPrebuiltAar is deriving from AndroidLibrary so that it can
propagate classpaths and such like it. This means that it needs to
create a JarBuildStepsFactory. That factory isn't actually used because
AndroidPrebuiltAar doesn't expose any of its own SourcePath outputs to
dependents.

Make this simpler and not pass totally unused things into the created
factory.

Reviewed By: styurin

fbshipit-source-id: e2d9469d03