Precompute and cache hashcode for AbstractBuildRule
Summary:
Cache the hashcode for BuildRules.
Even though this delegates to BuildTarget's hashcode, which is cached,
jvm overhead of method call, and how frequently this is used makes
about 5-10% time savings on AG construction and 5% savings on parsing.

This is about 10-15% improvement on newer guava versions due to different
set implementation.

Reviewed By: ttsugriy

fbshipit-source-id: 686dd78496