Fix ImmutableMapWithNullValues to not create a set of MapEntries
Summary:
MapEntry have a hashcode of 0 when key == entry.
We created sets of MapEntries for some of our implementation, which
would result in numerous hash collisions and poor performance,
especially in new guava versions due to hash flooding detection that
delegates back to jdk hashset.

This changes delegates to specialized MapEntry sets that do not suffer
this issue.

Reviewed By: sbalabanov

fbshipit-source-id: afda2f46c1