SHA256 Patch to V2 Signing
Summary:
After we migrate local build of Android apps to v2 signing, we at the same time switched v1 signing algorithm from SHA1 to SHA256, which provides much stronger security guarantee.

However, AgentUtil which checks for Manifest Digest(v1 hash) only looks for SHA1 hashing, and would throw an error if the APK is signed with SHA256. This patch instructs AgentUtil to look for both algorithms of hashing.

The changes in this diff follow the instruction in assets/android/README.txt.

Reviewed By: styurin

fbshipit-source-id: 473de98562