allow cgo_library to pull go native dependencies + allow to use standard (non-cgo) sources
Summary:
In OSS go there is the cgo package and go package are one being. This means that you can have source files that are pure .go (without import "C") and cgo (with import "C").
Go build would scan the sources and split them into two groups:
1. go files
2. cgo files

In order to provide the same functionality (both go and cgo files in one package) we need to specify by hand which file is what, ```go_srcs``` attribute. Following this scheme we need to allow the go native deps to be passed as well.

Reviewed By: philipjameson

fbshipit-source-id: 25c512fe6a