Clang on windows cxx toolchain implementation.
Summary:
This implements flavors of Compiler, Preprocessor, and Archiver for the Clang on Windows toolchain.
Notably, this does not specify clang _targeting_ Windows (for the Compiler implementation),
it just overrides the default ClangCompiler with settings required to invoke Clang on Windows.
ClangWindowsArchiver, on the other hand, supports the LLVM thin libs extension, which we do
need to specially handle even if in all other respects it behaves the same as lib.exe.

This also implements ClangClCompiler, a subclass of WindowsCompiler, which has overrides
specific to clang-cl (which otherwise conforms to cl's interface).

The existing implementation of lld is compatible enough with link.exe that a separate class
is not needed to use it. Simply specifying the exe via cxx.ld= is enough. To enable lto,
extra flags can be provided via cxxflags and ldflags.

fbshipit-source-id: ff3b4aa406