add a more flexible template for linking to function topics
Summary:
Add a template for linking to function topics that enables you to link to //any// of the function topics.

Also, the new template enables you, optionally, to specify the text that is rendered for the link.

Until now, our template strategy has been to create a separate template for each function topic to which the author wants to link.

Reviewed By: ttsugriy

fbshipit-source-id: 4bd8da07b4