Better buckd death ceremony
Summary:
For POSIX systems, rely on process existence by checking pid rather then checking for socket file existence. On most POSIX implementations, closing the domain socket does not really remove the file, so waiting for 3 seconds for file to disappear is pointless. Instead we are checking for process to terminate, kill it hard if it does not, and then unlink the socket.
The process has not changed on Windows, though it should be, one day.

Reviewed By: ttsugriy

fbshipit-source-id: 5c5dc45f67