add declared_package_name to cgo_library
Summary:
As described in PR:
https://github.com/facebook/buck/issues/1999

The import path (in buck terminology package_name attribute) can be different from the package name declared in the the source files.

In the diff I modified the CGoGenImport to run GoList in order to find the package name directly from the .go file. Which file? the first one in the sorted list.

In theory all the files in the package should have the same package name but in buck we can put two files from the same directory into two different build targets. So the first file within defined source list is a better solution.

Reviewed By: styurin, nemith

fbshipit-source-id: 1de0236087