Introduce JarBuildStepsFactory.JavaDependencyInfo
Summary:
JarBuildStepsFactory had three fields that were derived from the
abi/compile-time classpaths. Logically, these three things are tightly
connected together, but because the factory takes them in as separate
things in the constructor it has lost that connection. Correctly adding
these to the rulekey is very difficult without being able to make that
assumption (e.g. the current factory doesn't add the
SourceOnlyAbiRuleInfoFactory to its rulekey and just depends on that
being properly reflected by other fields).

This introduces JavaDependencyInfo which holds all the information for a
single dependency that is required to understand how to derive its value
for abiClasspath/compileTimeClasspath/SourceOnlyAbiRuleInfo. These three
fields are then replaces by a single List<JavaDependencyInfo>.

Reviewed By: styurin

fbshipit-source-id: a1106d8d8a