Add support for @Value.Lazy/@Value.Derived
Summary:
Value.Lazy is really difficult to support for something that is
AddToRuleKey, and so we require that it be converted to Value.Derived
in that case.

Excluded Value.Lazy and all Value.Derived are pretty straightforward to
support, just need to update DefaultClassInfo to find the fields.

An error is thrown when encountering a lazy field that is added to the
rulekey. If the fields that are being used to compute that lazy value
are appropriately reflected in the rulekey, then the field can stay lazy
and just drop the AddToRuleKey.

Reviewed By: ttsugriy

fbshipit-source-id: 7be7d1cd4d