Don't copy Swift libraries to Framework targets
Summary:
Currently, Swift libraries are copied to all bundle targets except for app extensions. When a framework uses Swift, and then is embedded in an app, this leads to multiple copies of the standard libraries:

- One added to the framework.
- One added to the app itself.

We should only need the second one. It's possible that the correct thing to do here is:

```
boolean shouldCopySwiftStdlib = extension.equals(AppleBundleExtension.APP.toFileExtension())
```

But, I'm not sure about some of the other extensions - `PLUGIN`, `XPC`, `PREFPANE`, etc.

I have added a `.buckconfig` option to disable this if necessary: `swift.copy_stdlib_to_frameworks`.

Reviewed By: milend

fbshipit-source-id: 94707783d2