Rename ActionGraphCache to ActionGraphProvider
Summary:
`ActionGraphCache` is a gigantic class right now that contains a lot of logic. At the same time it's kept in `Daemon` which makes it's difficult to reason about the life cycle of this object and its dependencies.

The solution is to separate the cache storage and the logic around its creation. The storage will be implemented in a new `ActionGraphCache` class and the existing `ActionGraphCache` needs to be renamed to `ActionGraphProvider`.

Reviewed By: bobyangyf

fbshipit-source-id: 39e5d477c8