Tag Modularity application_module_blacklist as non deps
Summary:
Adding a module that had restricted visibility to the blacklist caused
a "module X is not vissible to Y" error. This happened due to the fact
that any BuildTarget present in the buck rule is assumed to be a dep.

This diff adds the hint that the buck target is not a dep.

Reviewed By: joelmccall

fbshipit-source-id: 1c2c99e33a