Do not sort include paths.
Summary:
A lot of CPU time is spent needlessly sorting included paths
even though they are only used for `buck audit includes` command.

As part of this, also swithced to storing paths as strings instead of
`Path`s to reduce memory usage and the number of conversions, since the
final format is a `String. anyways.

This change should help both parsers, but Skylark would benefit more.

Reviewed By: sbalabanov

fbshipit-source-id: a475d5271c