Added ability to specify an output name for CxxBinary.
Summary:
One of the consideration when implementing this is to make sure that users cannot override their outputs because of mistakes when specifying the same output name for different targets. This issue is resolved by making sure the output goes into buck-out/gen/<target_name>/<user specified output name>
From Buck's point of view, this change allows for targets in the same file to produce outputs with the same name (without this change Buck doesn't allow to have targets with same name in the same BUCK file, and the names of the outputs were always the target names, which prevents having targets produce binaries with the same name).

Reviewed By: cjhopman

fbshipit-source-id: bb1392de4a