Switch partial regen to traverse deps
Summary:
partial regen isn't really saving any time by not writing out deps, and it's actually costing us time when
we have to do `buck query` to find the first-order deps. This change switches the behavior of partial regen to write
the whole subtree of the requested module. Most of this should be a no-op now that the sha-check has been fixed.

Reviewed By: cute-jumper

fbshipit-source-id: 61a6b2a83a