Fix rule_key_diff command args issue
Summary: The consistency checker tool throws `IllegalAnnotationError` when the `rule_key_diff` subcommand is run. `index` on `Argument` defaults to 0, so we get a conflict for the target name argument. I could have sworn I ran this before and it worked, but it doesn't work now, and `args4j` hasn't been upgraded in a while, so \_(o_O)_/. No other subcommand has the same issue.

fbshipit-source-id: eff1fdfd43