Proposal for exposing current node for deps iteration
Summary:
Proposed solution for making it easier to use the 3-param version of `deps`.
Right now there's no way to e.g. traverse only the `provided_deps` edges of a node.

This diff implements a `lookup()` function for resolving general target variables, and
adds a `this` variable to represent the current target. This means that following `provided_deps` edges becomes:

`deps(top_level_target, labels("provided_deps", lookup(this)))`

Happy to accept any suggestions for naming of the `lookup()` function and the `this` variable

Reviewed By: tdrhq

fbshipit-source-id: 82042579be