Explicitly only follow native deps for java_test.use_cxx_libraries
Summary:
Right now, whether or not to include a native library in a java_test is
a very complicated computation that depends on the use of class abis or
source abis or source only abis and the particular dependency graph.

Change this to explicitly not traverse java dependencies to discover
native libraries to include.

Longer-term we should change this to use a packageable so that it makes
more sense.

Reviewed By: ttsugriy

fbshipit-source-id: bda5b81200