Fix broken link to (now non-existent) Apple docs for xcodebuild
Summary:
The documentation for `.buckconfig` has been linking to the Apple documentation for `xcodebuild`, but recently, Apple seems to have pulled that documentation from their site.

Fortunately, the `man` page for `xcodebuild` is truly excellent and is available after you install Xcode.

So: replace the broken link with instructions on how to find the same information in the `man` page.

Reviewed By: ttsugriy

fbshipit-source-id: 75375ddd5f