Improve error handling for invalid struct fields.
Summary:
Previously, in case an invalid field name is used, a cryptic
error message that does not include a full stack trace would be printed,
making it very hard to debug and fix the underlying problem.

Now, all field names are checks before they are used to construct a struct.

Reviewed By: philipjameson

fbshipit-source-id: e58a640f54