Add constraint_values to config_setting
Summary:
Hooking up constraints into `config_setting`.

`ConstraintResolver` is a new class that provides access constraints by `BuildTarget`. This is used by `ConfigSettingSelectable` to get access to constraints which are used to check if a platform matches the constraints.

Note that `RawTargetNodeToTargetNodeFactory` passes a `Platform` with an empty list of constraints. It means that non of the non-empty platforms would match it. The following changes will add passing `Platform` from CLI.

Reviewed By: bobyangyf

fbshipit-source-id: 40a5d83398