Use target platform from CLI to resolve selects
Summary:
This change allows to specify a list of target platforms through CLI. The first platform is used to resolve select that use `config_setting` with constraint values.

The name of the argument (`--target-platforms`) implies that multiple platforms can be specified, but only the first platform is used. This is intentional for the case we need to support a build with multiple target platforms in future.

Reviewed By: philipjameson

fbshipit-source-id: a26633e34f