add logic for tasks to be (optionally) cancelled on scheduling of repeat tasks
Summary: some tasks (e.g. trash cleaning) don't need to run multiple times if scheduled on top of each other. this enables clients to add a "should cancel" option to tasks. a task marked "should cancel" that has not yet run will be removed from the queue and not run if a second task with the same action is scheduled. if the task has already been submitted, it runs as usual. option to "hard cancel" maybe to add later.

Reviewed By: bobyangyf

fbshipit-source-id: b6f0416c68